AndroidTest04 complete
