+++ AndroidTest04/app/src/main/java/com/example/ryota/androidtest04/Profile.java
@@ -1,34 +1,22 @@
 package com.example.ryota.androidtest04;
 
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.util.Log;
 
-class Profile {
-    private final String name;
-    private final Integer age;
-    private final String sex;
-    private final String language;
-
+class Profile extends AppCompatActivity {
 
 
-    Profile(String name, Integer age, String sex, String language){
-        this.name = name;
-        this.age = age;
-        this.sex = sex;
-        this.language = language;
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        Account account1 = new Account("太郎",25,"男性","Swift");
+        Account account2 = new Account("花子",24,"女性","java");
+
+        account1.ProfileOutPut();
+        account2.ProfileOutPut();
+
     }
 
-    public String getName() {
-        return this.name;
-    }
 
-    public Integer getAge() {
-        return this.age;
-    }
-
-    public String getSex() {
-        return this.sex;
-    }
-
-    public String getLanguage() {
-        return this.language;
-    }
 }
